pid_serial_monitor/requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0
numpy>=1.21.0